Upgrade to latest version of DynamoDB Local
Amazon have stopped redirecting to an artifact and have instead opted
for a versioned bucket. Therefore, to continue to support reproducible
builds, we point to a specific version of the artifact. Assuming they
keep old versions around for a while, this should continue to work more
or less as expected.